Notification: Create NotificationContent_Bongo file
